fix(lint): split cleanup_tests.rs under the 500-line gate
